Add a Toast when Delete account

+++ app/src/main/java/com/example/todomateclone/ui/user/UserPageFragment.kt
@@ -85,11 +85,29 @@ class UserPageFragment : Fragment() {
         }
 
         deleteAccountButton.setOnClickListener() {
-            CoroutineScope(Dispatchers.IO).launch {
-                userViewModel.deleteUser()
-                userViewModel.logout()
-            }
-            this.findNavController().navigate(R.id.action_global_login_graph)
+            AlertDialog.Builder(mainActivity)
+                .setTitle("계정 삭제")
+                .setMessage("정말로 계정을 삭제하시겠습니까?\n연동 삭제 시 등록된 사용자 정보가 삭제되며 회원가입을 다시 진행해야합니다.\n앱은 자동 종료됩니다.")
+                .setPositiveButton("예"
+                ) { _, _ ->
+                    Log.d("MyTag", "positive")
+                    CoroutineScope(Dispatchers.IO).launch {
+                        userViewModel.deleteUser()
+                        userViewModel.logout()
+                        mainActivity.finishAffinity()
+                        System.runFinalization()
+                        exitProcess(0)
+                    }
+                    this.findNavController().navigate(R.id.action_global_login_graph)
+                }
+                .setNegativeButton("아니오"
+                ) { _, _ -> Log.d("MyTag", "negative") }
+                .setNeutralButton("돌아가기"
+                ) { _, _ -> Log.d("MyTag", "neutral") }
+                .create()
+                .show()
+
+
         }
 
         // 구글 계정 연동 해제
